requirements.txt
# Web Framework
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
gunicorn==21.2.0
Flask-SocketIO==5.3.6
python-socketio==5.10.0
python-engineio==4.8.0
bidict==0.22.1

# API and Data Processing
requests==2.31.0
urllib3==2.0.7
beautifulsoup4==4.12.2
lxml==4.9.3
jsonschema==4.19.1
tmdbv3api==1.7.7

# Data Science & ML
numpy==1.25.2
scipy==1.11.3
pandas==2.1.1
scikit-learn==1.3.1
nltk==3.8.1

# Speech Recognition & Synthesis
SpeechRecognition==3.10.0
pyttsx3==2.90
gTTS==2.3.2
pydub==0.25.1
sounddevice==0.4.6
PyAudio==0.2.13
vosk==0.3.45

# External APIs
deepgram-sdk==2.11.0

# Utilities
python-dotenv==1.0.0
pickleshare==0.7.5
pygame==2.5.2

# Development & Testing
pytest==7.4.2
black==23.9.1
flake8==6.1.0